% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcdist.R
\name{arcdist}
\alias{arcdist}
\title{Title}
\usage{
arcdist(lat, lon, lat1 = NULL, lon1 = NULL, scale = "nmi")
}
\arguments{
\item{lat}{XXX}

\item{lon}{XXX}

\item{lat1}{XXX}

\item{lon1}{XXX}

\item{scale}{XXX}
}
\description{
Title
}
