% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_calculate_distance.R
\name{vms_calculate_distance}
\alias{vms_calculate_distance}
\title{Calculates distance between pings}
\usage{
vms_calculate_distance(d)
}
\arguments{
\item{d}{A standardize vms dataframe}
}
\value{
A dataframe with a new variable distance (in units of nautical miles)
}
\description{
The function calculates the distance between consequtive pings for each
vessel.
}
