% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_calculate_duration.R
\name{vms_calculate_duration}
\alias{vms_calculate_duration}
\title{Calculates durations between pings}
\usage{
vms_calculate_duration(d)
}
\arguments{
\item{d}{A standardize vms dataframe}
}
\value{
A dataframe with a new variable duration (in units hours)
}
\description{
The function calculates the duration between consequtive pings for each
vessel.
}
