% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_calculate_speed.R
\name{vms_calculate_speed}
\alias{vms_calculate_speed}
\title{Calculates speed between pings}
\usage{
vms_calculate_speed(d)
}
\arguments{
\item{d}{A standardize vms dataframe}
}
\value{
A dataframe with a new variable speed (in units of nautical miles
per hour)
}
\description{
The function calculates the speed between consequtive pings for each
vessel. If either a variable distance or duration or both is not available
it gets calculated from the data.
}
