% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_categorise_fishing.R
\name{vms_categorise_fishing}
\alias{vms_categorise_fishing}
\title{Categorizes fishing activity based on vessel speed}
\usage{
vms_categorise_fishing(d, fishing.lim)
}
\arguments{
\item{d}{A standardized vms dataframe}

\item{fishing.lim}{A vector specifying the lower and uppper value of
instantaneous vessel speed that is classified as fishing}
}
\value{
A dataframe with
}
\description{
Creates two new variables, "activity" indicating speed classification and
"fishing", a boolean variable where TRUE is considered fishing and FALSE
not fishing.
}
