% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_compute_effort.R
\name{vms_compute_effort}
\alias{vms_compute_effort}
\title{Title}
\usage{
vms_compute_effort(d, res, lon.min = -179, lon.max = 179, lat.min = -89,
  lat.max = 89)
}
\arguments{
\item{d}{XXX}

\item{res}{XXX}

\item{lon.min}{XXX}

\item{lon.max}{XXX}

\item{lat.min}{XXX}

\item{lat.max}{XXX}
}
\description{
Title
}
