% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_convert_to_time.R
\name{vms_convert_to_time}
\alias{vms_convert_to_time}
\title{Title}
\usage{
vms_convert_to_time(x)
}
\arguments{
\item{x}{Character vector representing time}
}
\value{
Time vector
}
\description{
Title
}
