% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_data_summary.R
\name{vms_data_summary}
\alias{vms_data_summary}
\title{Summarise vms data}
\usage{
vms_data_summary(d)
}
\arguments{
\item{d}{A standardized vms dataframe}
}
\value{
A dataframe
}
\description{
Provides some summary statistics on key variables of the imported and
standarized vms data.
}
