% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_export_tracks.R
\name{vms_export_tracks}
\alias{vms_export_tracks}
\title{Title}
\usage{
vms_export_tracks(d, file.name = "data-products/Vessel_tracks")
}
\arguments{
\item{d}{Standardized vms dataframe}

\item{file.name}{A character vector that}
}
\description{
Title
}
