% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_filter_data.R
\name{vms_filter_data}
\alias{vms_filter_data}
\title{Filter standardized vms dataframe}
\usage{
vms_filter_data(d, lon.lim, lat.lim, date.lim, speed.lim)
}
\arguments{
\item{d}{Standardized vms data frame}

\item{lon.lim}{A vector of two value, the lower and upper longitudinal range
to be retained.}

\item{lat.lim}{A vector of two value, the lower and upper latitudinal range
to be retained.}

\item{date.lim}{A vector of two value, the lower and upper date range
to be retained. The vector has to be on the format of c("yyyy-mm-dd hh:mm:ss",
"yyyy-mm-dd hh:mm:ss")}

\item{speed.lim}{A vector of two value, the lower and upper speed range
to be retained.}
}
\value{
A dataframe
}
\description{
Take a standarized vms dataframe and filters data not within any of the
specified boundarise of longitude, latitude, date and speed. The user can
specify the filter on one variable or all. Those not specified will not be
affected.
}
