% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_import_data2.R
\name{vms_import_data2}
\alias{vms_import_data2}
\title{Standardize vms data}
\usage{
vms_import_data2(file.name, country)
}
\arguments{
\item{file.name}{XXX}

\item{country}{XXX}
}
\value{
dataframe
}
\description{
XXX
}
