% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_plot_speed2.R
\name{vms_plot_speed2}
\alias{vms_plot_speed2}
\title{Histogram of speed limits}
\usage{
vms_plot_speed2(d, speed.min = 2, speed.max = 4.5, speed.limit = 15,
  bindwidth = 0.5)
}
\arguments{
\item{d}{A standardized VMS data frame}

\item{speed.min}{XXX}

\item{speed.max}{XXX}

\item{speed.limit}{A value that indicates maximum plausable speed. Any speed
higher than this valule will not be plotted}

\item{bindwidth}{The binwidth of the histogram}
}
\value{
dataframe
}
\description{
A wrapper for a histogram plot. It is expected that the dataframe contains variable labelled "speed". User can control the bindwidth and cut off
inplausable speed in the plot
}
