% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_plot_track.R
\name{vms_plot_track}
\alias{vms_plot_track}
\title{Title}
\usage{
vms_plot_track(d, lon.lim, lat.lim)
}
\arguments{
\item{d}{A standardized vms dataframe}

\item{lon.lim}{A vector specifying the lower and upper longitude boundaries
to be plotted.}

\item{lat.lim}{A vector specifying the lower and upper latitudinal boundaries
to be plotted.}
}
\description{
Title
}
