% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_setup_directory.R
\name{vms_setup_directory}
\alias{vms_setup_directory}
\title{Create directories}
\usage{
vms_setup_directory(example = TRUE)
}
\arguments{
\item{example}{Boolean, if TRUE (defaul) will copy the
\url{ftp://ftp.hafro.is/pub/reiknid/einar/vms/iceland_vms.csv} into the
data-raw directory.}
}
\description{
The function creates three directories in the current working
directory: data-raw, data-tidy and data-product. Also seeds in example
VMS data set upon request (requires internet connection).
}
