% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_setup_templates.R
\name{vms_setup_templates}
\alias{vms_setup_templates}
\title{Seeds three rmarkdown documents}
\usage{
vms_setup_templates()
}
\description{
The function installs three rmarkdown documents in the current
working directory: 01vms_cleanup.Rmd, 02vms_rasterize.Rmd and 03vms_tracks.Rmd.
The default specification in the rmarkdown documents are set in order to
run VMS example data that reside in the data-raw directory
(see \code{\link{vms_setup_directory}}).
}
