% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_standardize_names.R
\name{vms_standardize_names}
\alias{vms_standardize_names}
\title{Title}
\usage{
vms_standardize_names(cn)
}
\arguments{
\item{cn}{A character vector}
}
\value{
A character vector
}
\description{
Title
}
