% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_vessel_summary.R
\name{vms_vessel_summary}
\alias{vms_vessel_summary}
\title{Summarise vms data by vessel}
\usage{
vms_vessel_summary(d)
}
\arguments{
\item{d}{A standardized vms dataframe}
}
\value{
A dataframe
}
\description{
Provides some summary statistics on key variables of the imported and
standarized vms data by vessel
}
