% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_import_data.R
\name{vms_import_data}
\alias{vms_import_data}
\title{Standardize vms data}
\usage{
vms_import_data(file.name, country, lon.min, lon.max, lat.min, lat.max)
}
\arguments{
\item{file.name}{XXX}

\item{country}{XXX}

\item{lon.min}{XXX}

\item{lon.max}{XXX}

\item{lat.min}{XXX}

\item{lat.max}{XXX}
}
\value{
dataframe
}
\description{
XXX
}
