% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_plot_raster.R
\name{vms_plot_raster}
\alias{vms_plot_raster}
\title{Title}
\usage{
vms_plot_raster(d, lon.lim, lat.lim)
}
\arguments{
\item{d}{A tidy vms dataframe containing variables "lon" and "lat"}

\item{lon.lim}{A vector containing lower and upper longitudes to plot}

\item{lat.lim}{A vector containing lower and upper latitudes to plot}
}
\description{
Title
}
