% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_plot_speed.R
\name{vms_plot_speed}
\alias{vms_plot_speed}
\title{Histogram of speed limits}
\usage{
vms_plot_speed(d, bindwidth = 0.5, speed.limit = 15)
}
\arguments{
\item{d}{A standardized VMS data frame}

\item{bindwidth}{The binwidth of the histogram}

\item{speed.limit}{A value that indicates maximum plausable speed. Any speed
higher than this valule will not be plotted}
}
\value{
dataframe
}
\description{
A wrapper for a histogram plot. It is expected that the dataframe contains variable labelled "speed". User can control the bindwidth and cut off
inplausable speed in the plot
}
